refac(UploadForms): separate forms Separated FileUploadForm and EmptyUploadForm, as well as tests for each
