+++ Frontend/src/MainPage/Uploads/FileUploadForm.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { screen, render, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 
-import { FileUploadForm, EmptyUploadForm } from './UploadForms';
+import FileUploadForm from './FileUploadForm';
 
 
 describe('FileUploadForm', () => {
@@ -214,63 +214,6 @@ describe('FileUploadForm', () => {
 
   {/* passes in CSRF token -- how to test? */}
   // test('passes in CSRF token', () => {
-    
-  // });
-});
-
-
-describe('EmptyUploadForm', () => {
-  const setFiles = (newFiles => {appState.files = [...newFiles]});
-  const appState = {files: [], setFiles: setFiles};
-  const handleInput = jest.fn();
-
-  test('renders upload button with icon', () => {
-    render(
-      <EmptyUploadForm
-        files={appState.files}
-        setFiles={appState.setFiles}
-        handleInput={handleInput}
-        uploadResult={'Success'}
-      />
-    );
-
-    expect(screen.getByText('Upload Here!')).toBeInTheDocument();
-    expect(screen.getByTestId('UploadIcon')).toBeInTheDocument();
-  });
-
-  test('includes file input', () => {
-    render(
-      <EmptyUploadForm
-        files={appState.files}
-        setFiles={appState.setFiles}
-        handleInput={handleInput}
-        uploadResult={'Success'}
-      />
-    );
-
-    expect(screen.getByTestId('fileInput')).toBeInTheDocument();
-  });
-
-  test('Shows error when applicable', () => {
-    render(
-      <EmptyUploadForm
-        files={appState.files}
-        setFiles={appState.setFiles}
-        handleInput={handleInput}
-        uploadResult={'Why is the rum always gone?'}
-      />
-    );
-
-    expect(screen.getByRole('alert')).toBeInTheDocument();
-    expect(screen.getByText(/rum/)).toBeInTheDocument();
-  });
-
-  // /* vvvvv how to test these? vvvvv */
-  // test('input allows file selection', () => {
-    
-  // });
-
-  // test('input accepts files', () => {
     
   // });
 });
